requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
scikit-learn>=1.2.0
pyyaml>=6.0
tqdm>=4.65.0
huggingface-hub>=0.16.0
numpy>=1.23.0
